+++ app/map-components/custom-marker-ng-for.component.ts
@@ -1,28 +1,22 @@
 import { Component } from '@angular/core';
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs/Observable';
+import { SourceCodeService } from '../source-code.service';
 import 'rxjs/add/observable/of';
 
-let templateStr = `
-  <h1>Custom Marker With *ngFor</h1>
-  <ng2-map zoom="13" center="Brampton, Canada">
-    <custom-marker *ngFor="let pos of positions" [position]="pos">
-      <div class="custom-icon">{{count}}</div>
-    </custom-marker>
-  </ng2-map>
-  <button (click)="positions = getRandomMarkers()">Show Random Markers</button> <br/>
-  <button (click)="showMarkersFromObservable()">Show Random Markers From Observable</button> <br/>
-  <button (click)="count = count + 1">Increment</button> <br/>
-  <code>
-   <br/><b>HTML</b>
-    <pre>{{templateStr | htmlCode:'-code'}}</pre>
-    <b>function getRandomMarkers</b> 
-    <pre>{{getRandomMarkers | jsCode}}</pre>
-    <b>function showMarkersFromObservable</b> 
-    <pre>{{showMarkersFromObservable | jsCode}}</pre>
-  </code>
-`;
 @Component({
-  template: templateStr,
+  template: `
+    <h1>Custom Marker With *ngFor</h1>
+    <ngui-map zoom="13" center="Brampton, Canada">
+      <custom-marker *ngFor="let pos of positions" [position]="pos">
+        <div class="custom-icon">{{count}}</div>
+      </custom-marker>
+    </ngui-map>
+    <button (click)="positions = getRandomMarkers()">Show Random Markers</button> <br/>
+    <button (click)="showMarkersFromObservable()">Show Random Markers From Observable</button> <br/>
+    <button (click)="count = count + 1">Increment</button> <br/>
+    <button (click)="sc.plnkr(code)">See in plunker</button>
+    <pre class="prettyprint">{{code}}</pre>
+  `,
   styles: [
     `
       .custom-icon {
@@ -41,10 +35,11 @@ let templateStr = `
 export class CustomMarkerNgForComponent {
   public positions= [];
   public count: number = 0;
+  public code: string;
 
-  templateStr: string = templateStr;
-  constructor() {
+  constructor(public sc: SourceCodeService) {
     this.positions = this.getRandomMarkers();
+    sc.getText('CustomMarkerNgForComponent').subscribe(text => this.code = text);
   }
 
   getRandomMarkers() {
@@ -63,6 +58,6 @@ export class CustomMarkerNgForComponent {
     Observable.of(this.getRandomMarkers()) // Think this as http call
       .subscribe( positions => {
         this.positions = positions;
-      })
+      });
   }
 }