+++ app/map-components/custom-marker.component.ts
@@ -1,22 +1,26 @@
-import { Component, ElementRef } from '@angular/core';
+import { Component } from '@angular/core';
+import { SourceCodeService } from '../source-code.service';
 
-let templateStr: string = `
-  <h1>Custom Marker</h1>
-  <ng2-map center="Brampton, Canada">
-    <custom-marker position="Brampton, Canada">
-      <div><b>Hi, USA</b>
-        <img src="http://icons.iconarchive.com/icons/custom-icon-design/2014-world-cup-flags/32/USA-icon.png" />
-      </div>
-    </custom-marker>
-    <marker position="Brampton, Canada"></marker>
-  </ng2-map>
-  <code>
-    <br/><b>HTML</b>
-    <pre>{{templateStr | htmlCode:'-code'}}</pre>    
-  </code>
-`;
+@Component({
+  template: `
+    <h1>Custom Marker</h1>
+    <ngui-map center="Brampton, Canada">
+      <custom-marker position="Brampton, Canada">
+        <div><b>Hi, USA</b>
+          <img src="http://icons.iconarchive.com/icons/custom-icon-design/2014-world-cup-flags/32/USA-icon.png" />
+        </div>
+      </custom-marker>
+      <marker position="Brampton, Canada"></marker>
+    </ngui-map>
 
-@Component({ template: templateStr })
+    <button (click)="sc.plnkr(code)">See in plunker</button>
+
+    <pre class="prettyprint">{{code}}</pre>
+  `})
+
 export class CustomMarkerComponent {
-  templateStr: string = templateStr;
+  code: string;
+  constructor(public sc: SourceCodeService) {
+    sc.getText('CustomMarkerComponent').subscribe(text => this.code = text);
+  }
 }