+++ app/map-components/simple-info-window.component.ts
@@ -1,32 +1,44 @@
 import { Component } from '@angular/core';
+import { SourceCodeService } from '../source-code.service';
 
-let templateStr: string = `
-  <h1>Simple InfoWindow</h1>
-  <ng2-map center="Brampton, Canada">
-    <marker position="Brampton, Canada" draggable="true" (click)="clicked($event)"></marker>
-    <info-window id="iw">
-      lat: [[lat]], lng: [[lng]]
-    </info-window>
-  </ng2-map>
-  Please click the marker to see a info window
-  
-  <code>
-    <br/><b>HTML</b>
-    <pre>{{templateStr | htmlCode:'-code'}}</pre>
-    
-    <b>function clicked</b> 
-    <pre>{{clicked | jsCode}}</pre>
-  </code>
-  `;
 @Component({
-  template: templateStr
-})
+  template: `
+    <h1>Simple InfoWindow</h1>
+    <ngui-map center="Brampton, Canada">
+      <marker position="Brampton, Canada" draggable="true" (click)="clicked($event)"></marker>
+      <info-window id="iw">
+        <div *ngIf="marker.display">
+          lat: {{ marker.lat }}, lng: {{ marker.lng }}
+        </div>
+        <button (click)="hideMarkerInfo()">Hide Info</button>
+      </info-window>
+    </ngui-map>
+    Please click the marker to see a info window
+
+    <button (click)="sc.plnkr(code)">See in plunker</button>
+
+    <pre class="prettyprint">{{code}}</pre>
+    `})
 export class SimpleInfoWindowComponent {
-  templateStr: string = templateStr;
-  clicked(event) {
-    let marker = event.target;
-    marker.ng2MapComponent.openInfoWindow('iw', marker, {
-      lat: marker.getPosition().lat(), lng: marker.getPosition().lng(),
-    });
+  marker = {
+    display: true,
+    lat: null,
+    lng: null,
+  };
+  code: string;
+
+  constructor(public sc: SourceCodeService) {
+    sc.getText('SimpleInfoWindowComponent').subscribe(text => this.code = text);
+  }
+
+  clicked({target: marker}) {
+    this.marker.lat = marker.getPosition().lat();
+    this.marker.lng = marker.getPosition().lng();
+
+    marker.nguiMapComponent.openInfoWindow('iw', marker);
+  }
+
+  hideMarkerInfo() {
+    this.marker.display = !this.marker.display;
   }
 }