+++ src/components/info-window.ts
@@ -1,19 +1,15 @@
 import {
   Component,
   ElementRef,
-  OnInit,
-  OnChanges,
-  OnDestroy,
   EventEmitter,
   SimpleChanges,
+  ViewChild, ViewContainerRef,
+  Output, OnInit, OnChanges, OnDestroy
 } from '@angular/core';
 import { Subject } from 'rxjs/Subject';
-import 'rxjs/add/operator/debounceTime';
-
-import { OptionBuilder } from '../services/option-builder';
-import { Ng2Map } from '../services/ng2-map';
-import { Ng2MapComponent } from './ng2-map.component';
-
+import { debounceTime } from 'rxjs/operator/debounceTime';
+import { NguiMap } from '../services/ngui-map';
+import { NguiMapComponent } from './ngui-map.component';
 
 const INPUTS = [
   'content', 'disableAutoPan', 'maxWidth', 'pixelOffset', 'position', 'zIndex', 'options'
@@ -23,24 +19,23 @@ const OUTPUTS = [
 ];
 
 @Component({
-  selector: 'ng2-map > info-window',
+  selector: 'ngui-map > info-window',
   inputs: INPUTS,
   outputs: OUTPUTS,
-  template: `<ng-content></ng-content>`,
+  template: `<div #template><ng-content></ng-content></div>`,
 })
 export class InfoWindow implements OnInit, OnChanges, OnDestroy {
-  public el: HTMLElement;
+  @Output() initialized$: EventEmitter<any> = new EventEmitter();
+
   public infoWindow: google.maps.InfoWindow;
   public objectOptions: google.maps.InfoWindowOptions = {};
   public inputChanges$ = new Subject();
-
-  public template: string;
-  public initialized$: EventEmitter<any> = new EventEmitter();
+  @ViewChild('template', {read: ViewContainerRef}) template: ViewContainerRef;
 
   constructor(
-    private ng2MapComponent: Ng2MapComponent,
     private elementRef: ElementRef,
-    private ng2Map: Ng2Map
+    private nguiMap: NguiMap,
+    private nguiMapComponent: NguiMapComponent,
   ) {
     this.elementRef.nativeElement.style.display = 'none';
     OUTPUTS.forEach(output => this[output] = new EventEmitter());
@@ -48,10 +43,10 @@ export class InfoWindow implements OnInit, OnChanges, OnDestroy {
 
   // Initialize this map object when map is ready
   ngOnInit() {
-    if (this.ng2MapComponent.mapIdledOnce) { // map is ready already
+    if (this.nguiMapComponent.mapIdledOnce) { // map is ready already
       this.initialize();
     } else {
-      this.ng2MapComponent.mapReady$.subscribe(map => this.initialize());
+      this.nguiMapComponent.mapReady$.subscribe(map => this.initialize());
     }
   }
 
@@ -62,49 +57,44 @@ export class InfoWindow implements OnInit, OnChanges, OnDestroy {
   // called when map is ready
   initialize(): void {
     console.log('infowindow is being initialized');
-    this.template = this.elementRef.nativeElement.innerHTML;
 
-    this.objectOptions = this.ng2MapComponent.optionBuilder.googlizeAllInputs(INPUTS, this);
+    this.objectOptions = this.nguiMapComponent.optionBuilder.googlizeAllInputs(INPUTS, this);
     this.infoWindow = new google.maps.InfoWindow(this.objectOptions);
     this.infoWindow['mapObjectName'] = 'InfoWindow';
     console.log('INFOWINDOW objectOptions', this.objectOptions);
 
-    // register infoWindow ids to Ng2Map, so that it can be opened by id
-    this.el = this.elementRef.nativeElement;
-    if (this.el.id) {
-      this.ng2MapComponent.infoWindows[this.el.id] = this;
+    // register infoWindow ids to NguiMap, so that it can be opened by id
+    if (this.elementRef.nativeElement.id) {
+      this.nguiMapComponent.infoWindows[this.elementRef.nativeElement.id] = this;
     } else {
       console.error('An InfoWindow must have an id. e.g. id="detail"');
     }
 
     // set google events listeners and emits to this outputs listeners
-    this.ng2Map.setObjectEvents(OUTPUTS, this, 'infoWindow');
+    this.nguiMap.setObjectEvents(OUTPUTS, this, 'infoWindow');
 
     // update object when input changes
-    this.inputChanges$
-      .debounceTime(1000)
-      .subscribe((changes: SimpleChanges) => this.ng2Map.updateGoogleObject(this.infoWindow, changes));
+    debounceTime.call(this.inputChanges$, 1000)
+      .subscribe((changes: SimpleChanges) => this.nguiMap.updateGoogleObject(this.infoWindow, changes));
 
-    this.ng2MapComponent.addToMapObjectGroup('InfoWindow', this.infoWindow);
+    this.nguiMapComponent.addToMapObjectGroup('InfoWindow', this.infoWindow);
     this.initialized$.emit(this.infoWindow);
   }
 
-  open(anchor: google.maps.MVCObject, data: any) {
-    let html = this.template;
-
-    for (let key in data) {
-      this[key] = data[key];
-      html = html.replace(`[[${key}]]`, data[key]);
-    }
-
+  open(anchor: google.maps.MVCObject) {
     // set content and open it
-    this.infoWindow.setContent(html);
-    this.infoWindow.open(this.ng2MapComponent.map, anchor);
+    this.infoWindow.setContent(this.template.element.nativeElement);
+    this.infoWindow.open(this.nguiMapComponent.map, anchor);
   }
-
+  close() {
+    // check if infoWindow exists, and closes it
+    if (this.infoWindow)
+      this.infoWindow.close();
+  }
   ngOnDestroy() {
+    this.inputChanges$.complete();
     if (this.infoWindow) {
-      OUTPUTS.forEach(output => google.maps.event.clearListeners(this.infoWindow, output));
+      this.nguiMap.clearObjectEvents(OUTPUTS, this, 'infoWindow');
       delete this.infoWindow;
     }
   }