+++ src/directives/base-map-directive.ts
@@ -1,28 +1,29 @@
-import { EventEmitter, SimpleChanges, OnInit, OnChanges, OnDestroy, Output } from '@angular/core';
+import {EventEmitter, SimpleChanges, Output, OnInit, OnChanges, OnDestroy} from '@angular/core';
 
 import { OptionBuilder } from '../services/option-builder';
-import { Ng2Map } from '../services/ng2-map';
-import { Ng2MapComponent } from '../components/ng2-map.component';
-
+import { NguiMap } from '../services/ngui-map';
+import { NguiMapComponent } from '../components/ngui-map.component';
+import { missingLibraryError } from '../services/util';
 export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
-  @Output() public initialized$: EventEmitter<any> = new EventEmitter();
+  // this should be redefined on each childr directive
+  @Output() initialized$: EventEmitter<any> = new EventEmitter();
 
   public mapObject: any; // e.g. google.maps.Marker
   public objectOptions: any; // e.g. google.maps.MarkerOptions
 
-  public ng2Map: Ng2Map;
+  public nguiMap: NguiMap;
   public optionBuilder: OptionBuilder;
   public libraryName: string;
   protected _subscriptions = [];
 
   constructor(
-    protected ng2MapComponent: Ng2MapComponent,
+    protected nguiMapComponent: NguiMapComponent,
     public    mapObjectName: string,
     protected inputs: string[],
     protected outputs: string[]
   ) {
-    this.ng2Map = this.ng2MapComponent['ng2Map'];
-    this.optionBuilder = this.ng2MapComponent['optionBuilder'];
+    this.nguiMap = this.nguiMapComponent['nguiMap'];
+    this.optionBuilder = this.nguiMapComponent['optionBuilder'];
     // all outputs must be initialized
     this.outputs.forEach(output => this[output] = new EventEmitter());
     this.mapObjectName = mapObjectName;
@@ -30,10 +31,10 @@ export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
 
   // Initialize this map object when map is ready
   ngOnInit() {
-    if (this.ng2MapComponent.mapIdledOnce) { // map is ready already
+    if (this.nguiMapComponent.mapIdledOnce) { // map is ready already
       this.initialize();
     } else {
-      this.ng2MapComponent.mapReady$.subscribe(map => this.initialize());
+      this.nguiMapComponent.mapReady$.subscribe(map => this.initialize());
     }
   }
 
@@ -48,18 +49,21 @@ export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
 
     // noinspection TypeScriptUnresolvedFunction
     if (this.libraryName) {
+      if (!google.maps[this.libraryName]) {
+        throw missingLibraryError(this.mapObjectName, this.libraryName);
+      }
       this.mapObject = new google.maps[this.libraryName][this.mapObjectName](this.objectOptions);
     } else {
       this.mapObject = new google.maps[this.mapObjectName](this.objectOptions);
     }
-    this.mapObject.setMap(this.ng2MapComponent.map);
+    this.mapObject.setMap(this.nguiMapComponent.map);
     this.mapObject['mapObjectName'] = this.mapObjectName;
-    this.mapObject['ng2MapComponent'] = this.ng2MapComponent;
+    this.mapObject['nguiMapComponent'] = this.nguiMapComponent;
 
     // set google events listeners and emits to this outputs listeners
-    this.ng2Map.setObjectEvents(this.outputs, this, 'mapObject');
+    this.nguiMap.setObjectEvents(this.outputs, this, 'mapObject');
 
-    this.ng2MapComponent.addToMapObjectGroup(this.mapObjectName, this.mapObject);
+    this.nguiMapComponent.addToMapObjectGroup(this.mapObjectName, this.mapObject);
     this.initialized$.emit(this.mapObject);
   }
 
@@ -67,18 +71,16 @@ export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
   // e.g., when map center is changed by user, update center on the map
   ngOnChanges(changes: SimpleChanges) {
     console.log(this.mapObjectName, 'objectOptions are changed', changes);
-    this.ng2Map.updateGoogleObject(this.mapObject, changes);
+    this.nguiMap.updateGoogleObject(this.mapObject, changes);
   }
 
   // When destroyed, remove event listener, and delete this object to prevent memory leak
   ngOnDestroy() {
     this._subscriptions.map(subscription => subscription.unsubscribe());
-    this.ng2MapComponent.removeFromMapObjectGroup(this.mapObjectName, this.mapObject);
+    this.nguiMapComponent.removeFromMapObjectGroup(this.mapObjectName, this.mapObject);
 
     if (this.mapObject) {
-      this.outputs.forEach(output => google.maps.event.clearListeners(this.mapObject, output));
-      this.mapObject['setMap'](null);
-      delete this.mapObject;
+      this.nguiMap.clearObjectEvents(this.outputs, this, 'mapObject');
     }
   }
 }