+++ src/directives/heatmap-layer.ts
@@ -1,21 +1,20 @@
 import { Directive } from '@angular/core';
 
 import { BaseMapDirective } from './base-map-directive';
-import { Ng2MapComponent } from '../components/ng2-map.component';
+import { NguiMapComponent } from '../components/ngui-map.component';
 
 const INPUTS = [ 'data', 'dissipating', 'gradient', 'maxIntensity', 'opacity', 'radius', 'options' ];
 const OUTPUTS = [];
 
 @Directive({
-  selector: 'ng2-map > heatmap-layer',
+  selector: 'ngui-map > heatmap-layer',
   inputs: INPUTS,
   outputs: OUTPUTS,
 })
 export class HeatmapLayer extends BaseMapDirective {
-  public libraryName: string;
+  public libraryName = 'visualization';
 
-  constructor(ng2MapComp: Ng2MapComponent) {
-    super(ng2MapComp, 'HeatmapLayer', INPUTS, OUTPUTS);
-    this.libraryName = 'visualization';
+  constructor(nguiMapComp: NguiMapComponent) {
+    super(nguiMapComp, 'HeatmapLayer', INPUTS, OUTPUTS);
   }
 }