+++ src/directives/kml-layer.ts
@@ -1,19 +1,18 @@
 import { Directive } from '@angular/core';
 
 import { BaseMapDirective } from './base-map-directive';
-import { Ng2MapComponent } from '../components/ng2-map.component';
+import { NguiMapComponent } from '../components/ngui-map.component';
 
 const INPUTS = [ 'clickable', 'preserveViewport', 'screenOverlays', 'suppressInfoWindows', 'url', 'zIndex', 'options' ];
 const OUTPUTS = [ 'click', 'defaultviewport_changed', 'status_changed' ];
 
 @Directive({
-  selector: 'ng2-map > kml-layer',
+  selector: 'ngui-map > kml-layer',
   inputs: INPUTS,
   outputs: OUTPUTS,
 })
 export class KmlLayer extends BaseMapDirective {
-
-  constructor(ng2MapComp: Ng2MapComponent) {
-    super(ng2MapComp, 'KmlLayer', INPUTS, OUTPUTS);
+  constructor(nguiMapComp: NguiMapComponent) {
+    super(nguiMapComp, 'KmlLayer', INPUTS, OUTPUTS);
   }
 }