+++ src/directives/places-auto-complete.ts
@@ -1,39 +1,47 @@
-import {Input, Output, Directive, EventEmitter, ElementRef, NgZone} from '@angular/core';
+import {
+  Input,
+  Output,
+  Directive,
+  EventEmitter,
+  ElementRef,
+} from '@angular/core';
 
+import { NgMapApiLoader } from '../services/api-loader';
 import { OptionBuilder } from '../services/option-builder';
-import { Ng2MapComponent } from '../components/ng2-map.component';
+import { missingLibraryError } from '../services/util';
 
 @Directive({
   selector: '[places-auto-complete]'
 })
-export class PlacesAutoComplete  {
-
+export class PlacesAutoComplete {
   @Input('bounds') bounds: any;
   @Input('componentRestrictions') componentRestrictions: any;
   @Input('types') types: string[];
 
   @Output('place_changed') place_changed: EventEmitter<any> = new EventEmitter();
-  @Output('initialized$')  initialized$: EventEmitter<any> = new EventEmitter();
+  @Output() initialized$: EventEmitter<any> = new EventEmitter();
 
   public objectOptions: any;
   public autocomplete: google.maps.places.Autocomplete;
 
   constructor(
     public optionBuilder: OptionBuilder,
-    public elementRef: ElementRef
+    public elementRef: ElementRef,
+    public apiLoader: NgMapApiLoader,
   ) {
-    if (typeof google === 'undefined' || typeof google.maps === 'undefined' || !google.maps.Map) {
-      this.addGoogleMapsApi();
-    } else {
-      this.initialize();
-    }
+    apiLoader.load();
+    apiLoader.api$.subscribe(() => this.initialize());
   }
 
   // only called when map is ready
-  initialize = ():void => {
+  initialize = (): void => {
     this.objectOptions =
       this.optionBuilder.googlizeAllInputs(['bounds', 'componentRestrictions', 'types'], this);
     console.log('places autocomplete options', this.objectOptions);
+
+    if (!google.maps.places) {
+      throw missingLibraryError('PlacesAutoComplete', 'places');
+    }
 
     this.autocomplete = new google.maps.places.Autocomplete(
       this.elementRef.nativeElement,
@@ -41,22 +49,10 @@ export class PlacesAutoComplete  {
     );
     console.log('this.autocomplete', this.autocomplete);
 
-    this.autocomplete.addListener('place_changed', place => this.place_changed.emit());
+    this.autocomplete.addListener('place_changed', place => {
+      this.place_changed.emit(this.autocomplete.getPlace());
+    });
 
     this.initialized$.emit(this.autocomplete);
-  };
-
-  addGoogleMapsApi(): void {
-    window['initializePlacesAutoComplete'] = this.initialize;
-    if ((!window['google'] || !window['google']['maps']) && !document.querySelector('#ng2-map-api')) {
-      let script = document.createElement( 'script' );
-      script.id = 'ng2-map-api';
-
-      // script.src = "https://maps.google.com/maps/api/js?callback=initNg2Map";
-      let apiUrl = Ng2MapComponent['apiUrl'] || 'https://maps.google.com/maps/api/js';
-      apiUrl += apiUrl.indexOf('?') ? '&' : '?';
-      script.src = apiUrl + 'callback=initializePlacesAutoComplete';
-      document.querySelector('body').appendChild(script);
-    }
   }
 }