+++ src/directives/polygon.ts
@@ -1,7 +1,7 @@
 import { Directive } from '@angular/core';
 
 import { BaseMapDirective } from './base-map-directive';
-import { Ng2MapComponent } from '../components/ng2-map.component';
+import { NguiMapComponent } from '../components/ngui-map.component';
 
 const INPUTS = [
   'clickable', 'draggable', 'editable', 'fillColor', 'fillOpacity', 'geodesic', 'paths',
@@ -13,13 +13,12 @@ const OUTPUTS = [
 ];
 
 @Directive({
-  selector: 'ng2-map>polygon, ng2-map>map-polygon',
+  selector: 'ngui-map>polygon, ngui-map>map-polygon',
   inputs: INPUTS,
   outputs: OUTPUTS,
 })
 export class Polygon extends BaseMapDirective {
-
-  constructor(ng2MapComp: Ng2MapComponent) {
-    super(ng2MapComp, 'Polygon', INPUTS, OUTPUTS);
+  constructor(nguiMapComp: NguiMapComponent) {
+    super(nguiMapComp, 'Polygon', INPUTS, OUTPUTS);
   }
 }