+++ src/directives/transit-layer.ts
@@ -1,18 +1,18 @@
 import { Directive } from '@angular/core';
 
 import { BaseMapDirective } from './base-map-directive';
-import { Ng2MapComponent } from '../components/ng2-map.component';
+import { NguiMapComponent } from '../components/ngui-map.component';
 
 const INPUTS = [];
 const OUTPUTS = [ ];
 
 @Directive({
-  selector: 'ng2-map > transit-layer',
+  selector: 'ngui-map > transit-layer',
   inputs: INPUTS,
   outputs: OUTPUTS,
 })
 export class TransitLayer extends BaseMapDirective {
-  constructor(ng2MapComp: Ng2MapComponent) {
-    super(ng2MapComp, 'TransitLayer', INPUTS, OUTPUTS);
+  constructor(nguiMapComp: NguiMapComponent) {
+    super(nguiMapComp, 'TransitLayer', INPUTS, OUTPUTS);
   }
 }