+++ src/services/geo-coder.ts
@@ -1,5 +1,7 @@
-import { Injectable } from '@angular/core';
-import { Subject } from 'rxjs/Subject';
+import { Injectable, OnDestroy } from '@angular/core';
+import { Observable } from 'rxjs/Observable';
+import { Observer } from 'rxjs/Observer';
+import { NgMapApiLoader } from './api-loader';
 
 /**
  *   Provides [defered/promise API](https://docs.angularjs.org/api/ng/service/$q)
@@ -7,22 +9,30 @@ import { Subject } from 'rxjs/Subject';
  */
 
 @Injectable()
-export class GeoCoder {
+export class GeoCoder implements OnDestroy {
+  private apiLoaderSubs = [];
+  constructor(private apiLoader: NgMapApiLoader) {}
 
   geocode(options: google.maps.GeocoderRequest) {
+    return new Observable((responseObserver: Observer<any>) => {
+        this.apiLoaderSubs.push(this.apiLoader.api$
+          .subscribe(() => this.requestGeocode(options, responseObserver)));
+    });
+  }
 
-    let geocode$ = new Subject();
-    let geocoder = new google.maps.Geocoder();
+  ngOnDestroy() {
+    this.apiLoaderSubs.map(sub => sub.unsubscribe());
+  }
 
+  private requestGeocode(options, observer) {
+    const geocoder = new google.maps.Geocoder();
     geocoder.geocode(options, function (results, status) {
       if (status === google.maps.GeocoderStatus.OK) {
-        geocode$.next(results);
+        observer.next(results);
+        observer.complete();
       } else {
-        geocode$.error(results);
+        observer.error(results);
       }
     });
-
-    return geocode$;
-  };
-
+  }
 }