+++ src/services/option-builder.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import { getJSON, IJson } from './util';
-import { GeoCoder } from './geo-coder';
 
 /**
  * change any object to google object options
@@ -8,8 +7,6 @@ import { GeoCoder } from './geo-coder';
  */
 @Injectable()
 export class OptionBuilder {
-
-  constructor(private geoCoder: GeoCoder) {}
 
   googlizeAllInputs(definedInputs: string[], userInputs: any) {
     let options: any = {};
@@ -49,7 +46,7 @@ export class OptionBuilder {
   googlize(input: any, options?: IJson): any {
     options = options || {};
     let output: any = input;
-    if (typeof input === 'string') { //convert string to a google object
+    if (typeof input === 'string') { // convert string to a google object
       if (input === 'false') {
         output = false;
       } else if (input === '0') {
@@ -72,26 +69,27 @@ export class OptionBuilder {
       }
     }
 
-    if (output instanceof Array) { //e.g., [1, 2]
-      if (options['key'] === 'bounds') {
-        output = new google.maps.LatLngBounds(output[0], output[1]);
-      }
-      else if (options['key'] === 'icons') {
-        output = this.getMapIcons(output);
-      }
-      else if (options['key'] === 'position' || (<string>options['key']).match(/^geoFallback/) ) {
-        output = this.getLatLng(output);
-      }
-    } else if (options['key'] && output instanceof Object) {
-      if (options['key'] === 'icon') {
-        output = this.getMarkerIcon(output);
-      }
-      else if ((<string>options['key']).match(/ControlOptions$/)) {
-        output = this.getMapControlOption(output);
+    if (options['key']) {
+      let key: string = <string>options['key'];
+      if (output instanceof Array) { // e.g., [1, 2]
+        if (key === 'bounds') {
+          output = new google.maps.LatLngBounds(output[0], output[1]);
+        } else if (key === 'icons') {
+          output = this.getMapIcons(output);
+        } else if (key === 'position' || key.match(/^geoFallback/) ) {
+          output = this.getLatLng(output);
+        }
+      } else if (output instanceof Object) {
+        if (key === 'icon') {
+          output = this.getMarkerIcon(output);
+        }
+        else if (key.match(/ControlOptions$/)) {
+          output = this.getMapControlOption(output);
+        }
       }
     }
 
-    //delete keys only for processing, not used by google
+    // delete keys only for processing, not used by google
     delete output['doNotConverStringToNumber'];
     delete output['key'];
 
@@ -135,8 +133,7 @@ export class OptionBuilder {
     let output: any;
     if (input.match(/^[A-Z][a-zA-Z0-9]+\(.*\)$/)) {
       try {
-        let exp = 'new google.maps.' + input;
-        output = Function(`return new google.maps.${input};`)()
+        output = Function(`return new google.maps.${input};`)();
       } catch (e) {}
     }
     return output;
@@ -239,7 +236,7 @@ export class OptionBuilder {
   }
 
   private onlyOptionsGiven(definedInputs: string[], userInputs: any): boolean {
-    for (let i = 0; i< definedInputs.length; i++) {
+    for (let i = 0; i < definedInputs.length; i++) {
       let input = definedInputs[i];
       if (input !== 'options' && typeof userInputs[input] !== 'undefined') {
         return false;