+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.17.0 (2017-03-29)
+
+  * renamed module to @ngui/map
+
+# 0.16.0 (2017-03-04)
+ 
+  * Added @Output initialized$ on all dierctives
+  * Fixed emit/subscribe issue, issue #99
+
 # 0.15.0 (2017-01-23)
 
   * added geo fallback options for map and marker
@@ -7,12 +16,14 @@
     * Fixed a bug of geocode callback losing setMethod name
     * Exposed `mapReady$` and `initialized$` as an Output
   * **0.15.3 (2017-02-13)**
-    * Made available for both imports of `Ng2MapModule` and `Ng2MapModule.forRoot()`
+    * Made available for both imports of `NguiMapModule` and `NguiMapModule.forRoot()`
   * **0.15.4 (2017-02-16)**
     * Upgraded webpack from 1.* tp 2.*
     * Fixed multiple map instance not loading on the first page.
-  * **0.15.5 (2017-02-16)**
-    * Removed debug line
+  * **0.15.5 (2017-02-16)** Removed debug line
+  * **0.15.6/7 (2017-02-16)**
+    * Fixed custom marker positioning
+    * issue #91, #92 fixed bug of not displaying map with places-autocomplete
 
 # 0.14.0 (2017-01-12)